feat：【PAY 支付】示例转账单，改成示例提现单，理解成本更低
